+++ Rest_Service/src/main/java/edu/BellevueCollege/NestedCatjam/ControlCognizant/ControlCognizantApplication.java
@@ -1,13 +1,13 @@
-package com.learningSpring.rest.webservices.restfulwebservices;
+package edu.BellevueCollege.NestedCatjam.ControlCognizant;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class RestfulWebServicesApplication {
+public class ControlCognizantApplication {
 
 	public static void main(String[] args) {
-		SpringApplication.run(RestfulWebServicesApplication.class, args);
+		SpringApplication.run(ControlCognizantApplication.class, args);
 	}
 
 }
