+++ Rest_Service/src/main/java/edu/BellevueCollege/NestedCatjam/ControlCognizant/Controllers/PostController.java
@@ -1,10 +1,10 @@
-package com.learningSpring.rest.webservices.restfulwebservices.Controllers;
+package edu.BellevueCollege.NestedCatjam.ControlCognizant.Controllers;
 
-import com.learningSpring.rest.webservices.restfulwebservices.Exceptions.UserNotFoundException;
-import com.learningSpring.rest.webservices.restfulwebservices.Entities.Post;
-import com.learningSpring.rest.webservices.restfulwebservices.Utils.PostDaoUtil;
-import com.learningSpring.rest.webservices.restfulwebservices.Entities.User;
-import com.learningSpring.rest.webservices.restfulwebservices.Utils.UserDaoUtil;
+import edu.BellevueCollege.NestedCatjam.ControlCognizant.Exceptions.UserNotFoundException;
+import edu.BellevueCollege.NestedCatjam.ControlCognizant.Entities.Post;
+import edu.BellevueCollege.NestedCatjam.ControlCognizant.Utils.PostDaoUtil;
+import edu.BellevueCollege.NestedCatjam.ControlCognizant.Entities.User;
+import edu.BellevueCollege.NestedCatjam.ControlCognizant.Utils.UserDaoUtil;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.bind.annotation.RestController;
