+++ Rest_Service/src/main/java/edu/BellevueCollege/NestedCatjam/ControlCognizant/Entities/Control.java
@@ -1,0 +1,21 @@
+package edu.BellevueCollege.NestedCatjam.ControlCognizant.Entities;
+
+import lombok.Data;
+import lombok.Getter;
+import lombok.Setter;
+
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.Id;
+import javax.persistence.Table;
+import java.util.UUID;
+
+@Entity
+@Data
+@Table(name = "control")
+public class Control {
+    @Id
+    @Column(name = "id", nullable = false)
+    private UUID id;
+
+}
