+++ Rest_Service/src/main/java/edu/BellevueCollege/NestedCatjam/ControlCognizant/Exceptions/UserNotFoundException.java
@@ -1,4 +1,4 @@
-package com.learningSpring.rest.webservices.restfulwebservices.Exceptions;
+package edu.BellevueCollege.NestedCatjam.ControlCognizant.Exceptions;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
